EIT-2484 | Implement JS and Golang interface (#14)
* Add JS SDK boilerplate

* Update gitignore

* Added js lib boilerplate

* Add js library boilerplate

* Initial record implementation

* Add record to go sdk

* Improve unwraping of config data

* Add header files and .a

* Begin proof proto api

* Finish proof server implementation

* Call proof proto API from Go SDK

* Make writer buffer bigger so proof response can fit

* Create Go client

* Handle error on record client call

* Add functions to create records

* Define alias types for the API

* Start implementing acceptance tests

* Add more acceptance tests and fix some broken ones

* Fix broken acceptance tests

* Add functional tests

* Begin implementation of JS interface

* Implement JS client

* Add functions to create records to JS client

* Add tests to JS client

* Fix jest config to accept ES modules

+++ bloock-core/src/record/entity/dto/record_write_response.rs
@@ -1,6 +1,6 @@
 #[derive(serde::Serialize, serde::Deserialize, Clone)]
 pub struct RecordWriteResponse {
-    pub anchor: i32,
+    pub anchor: i64,
     pub client: String,
     pub messages: Vec<String>,
     pub status: String,
