+++ bloock-core/src/record/entity/record_receipt.rs
@@ -2,7 +2,7 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq)]
 pub struct RecordReceipt {
-    pub anchor: i32,
+    pub anchor: i64,
     pub client: String,
     pub record: String,
     pub status: String,
